# shared/contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: Talk&Talk API
  version: 0.1.0
  description: |
    Frozen public contract for `/api/v1`.
    All JSON responses use a uniform envelope (see components.schemas.SuccessEnvelope / ErrorEnvelope).
    Breaking changes require `/api/v2` or coordinated release notes.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api-staging.talkandtalk.app/api/v1
    description: Staging
  - url: https://api.talkandtalk.app/api/v1
    description: Production
...
paths:
...
  /orders:
    get:
      tags: [orders]
      summary: List current user orders
      operationId: listOrders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      tags: [orders]
      summary: Create pending order
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companionId, themeId, durationMinutes, scheduledAt]
              properties:
                companionId: { type: string }
                serviceOfferingId:
                  type: string
                  minLength: 1
                  maxLength: 160
                  description: Optional active offering belonging to companionId. When present, its server-side price and duration are authoritative and frozen on the order.
                availabilityWindowId:
                  type: string
                  minLength: 1
                  maxLength: 160
                  description: Optional structured availability window returned by listCompanionAvailability. When supplied, scheduledAt must be a 30-minute candidate fully contained by this active window; the server records a capacity snapshot and rechecks it at confirmation and prepay.
                themeId: { type: string }
                durationMinutes:
                  type: integer
                  minimum: 30
                  maximum: 240
                  description: Must exactly match the selected service offering's duration when serviceOfferingId is supplied.
                scheduledAt: { type: string, format: date-time, description: Must fall within the server-configured commercial booking horizon. With availabilityWindowId it is the selected candidate start time and must use a 30-minute boundary. }
                clientRequestId:
                  type: string
                  minLength: 16
                  maxLength: 64
                  pattern: "^[A-Za-z0-9_-]+$"
                  description: Required in commercial release mode and stable across retries so one user action creates at most one order.
                recommendationImpressionId: { type: string, format: uuid, description: Optional card impression attribution, valid for 24 hours. }
      responses:
        "400":
          description: Invalid duration or schedule outside the allowed booking window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "200":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "409":
          description: Per-user or companion intake limit reached, an availability candidate/window is no longer valid or full, or an idempotency key was reused with different data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Commercial intake requires a clientRequestId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Order intake is paused or the controlled global open-order capacity is reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/service/today:
    get:
      tags: [orders, companions]
      summary: List the eligible companion owner's Beijing-today service checklist
      description: >-
        Returns only the caller's service entries whose appointment start falls in
        [00:00, 24:00) for Asia/Shanghai. It is a read-only workbench feed: no
        customer identity, conversation, refund, settlement, payment, or internal
        order fields are returned, and it never creates a schedule, reminder, or
        fulfillment action. pendingConfirmationCount is a separate count of the
        caller's still-unconfirmed pending requests; handling remains in the
        existing order flow.
      operationId: listMyCompanionTodayServiceChecklist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Today checklist ordered by scheduled appointment start
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CompanionTodayServiceSchedule"
        "403":
          description: The authenticated account is not an active identity-verified companion owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: No companion profile belongs to the authenticated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}:
    get:
      tags: [orders]
      summary: Order detail
      operationId: getOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /orders/{id}/timeline:
    get:
      tags: [orders]
      summary: List participant-safe order timeline events, including reschedule request facts when present
      operationId: listOrderTimeline
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Chronological order activity visible to the customer and the companion owner only
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [orderId, items]
                        properties:
                          orderId: { type: string }
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/OrderTimelineEvent"
        "404":
          description: Order not found or caller is not an order participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}/voice-room/access:
    post:
      tags: [orders]
      summary: Issue a short-lived, order-scoped real-time voice room credential
      description: Available only to the customer or companion owner of a voice SKU after manual acceptance, payment, and the companion's explicit service start. The server derives the room and participant identity; callers cannot choose either. UserSig and privateMapKey are short-lived response secrets and must not be persisted by clients. A refund, service completion, or elapsed service window can terminate an already connected room server-side; clients must handle the provider kicked-out event.
      operationId: issueOrderVoiceRoomAccess
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: A short-lived TRTC access credential for this order participant
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/VoiceRoomAccess"
        "404":
          description: Order not found or caller is not an order participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Order is not a voice service, is not accepted/started, is in after-sales handling, or its service window has ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Real-time voice has not been configured or the signing provider is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}/reschedule-requests:
    post:
      tags: [orders]
      summary: Submit a participant reschedule proposal without changing the current appointment
      description: The customer or companion owner may propose one future time. A structured order must supply an active structured availability window containing that candidate; the original order time, payment, refund and fulfillment state remain unchanged until a separate response accepts the proposal.
      operationId: createOrderRescheduleRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestedScheduledAt]
              properties:
                requestedScheduledAt:
                  type: string
                  format: date-time
                  description: Must be future, within the commercial booking horizon, and leave at least 15 minutes before a service can start.
                availabilityWindowId:
                  type: string
                  minLength: 1
                  maxLength: 160
                  description: Required when the original order uses structured availability. It must be an active companion window that contains the candidate and has current capacity.
      responses:
        "201":
          description: Pending proposal recorded and the other participant notified
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/OrderRescheduleRequest"
        "400":
          description: Requested time is invalid or outside the commercial booking horizon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Order state, current capacity, refund state, existing proposal, or response window prevents a new proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: A structured appointment did not include a new availability window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}/reschedule-requests/{requestId}/accept:
    post:
      tags: [orders]
      summary: Accept the other participant's reschedule proposal and atomically replace the booking slot
      description: The requester cannot accept their own proposal. The server locks the order, request and current candidate availability, then rechecks request expiry, refund state, appointment state and capacity before replacing the order's scheduled time and availability snapshot.
      operationId: acceptOrderRescheduleRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - name: requestId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Proposal accepted and order appointment snapshot atomically updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [rescheduleRequest, order]
                        properties:
                          rescheduleRequest:
                            $ref: "#/components/schemas/OrderRescheduleRequest"
                          order:
                            type: object
                            description: Updated order DTO, including its new scheduledAt and availabilitySnapshot.
        "403":
          description: The request creator attempted to accept their own proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Order or reschedule request not found, or caller is not an order participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Proposal expired, order changed state, a refund began, or the candidate is no longer available/capable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}/reschedule-requests/{requestId}/reject:
    post:
      tags: [orders]
      summary: Reject the other participant's reschedule proposal without changing the booking
      description: The requester cannot reject their own proposal. The server locks the order and request, confirms the caller is the other order participant, and records a rejection only while the proposal remains pending and valid. The order appointment, payment, refund and fulfillment state remain unchanged.
      operationId: rejectOrderRescheduleRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - name: requestId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Proposal rejected and its requester notified; the existing booking is unchanged
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/OrderRescheduleRequest"
        "403":
          description: The request creator attempted to reject their own proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Order or reschedule request not found, or caller is not an order participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Proposal has expired or is no longer pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}/cancel:
    post:
      tags: [orders]
      summary: Cancel pending/paying order
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /orders/{id}/completion-confirmations:
    post:
      tags: [orders]
      summary: Let the customer explicitly confirm a completed service without shortening the refund window
      operationId: confirmOrderCompletion
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Completion confirmation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "409":
          $ref: "#/components/responses/Error"
  /orders/{id}/service-guidelines-confirmations:
    post:
      tags: [orders]
      summary: Record one participant's pre-service acknowledgement of platform communication, service scope, and safety boundaries
      operationId: confirmOrderServiceGuidelines
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Role-specific service-guidelines acknowledgement recorded; it does not change payment, refund, or service-start eligibility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /orders/{id}/experience-feedback:
    post:
      tags: [orders]
      summary: Record one private customer experience-feedback submission for a completed order
      description: This is not a public companion review and cannot change order, refund, or settlement state. Free-text notes are moderated before storage; customers who need help use the existing order support-ticket flow.
      operationId: submitOrderExperienceFeedback
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                tags:
                  type: array
                  maxItems: 3
                  uniqueItems: true
                  items:
                    type: string
                    enum: [communicationClear, boundaryRespected, onTime, asExpected, needsImprovement]
                note: { type: string, maxLength: 200 }
      responses:
        "200":
          description: Private feedback recorded, or the original private feedback returned for a retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /orders/{id}/prepay:
    post:
      tags: [orders]
      summary: WeChat App or Mini Program prepay
      operationId: prepayOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                channel:
                  type: string
                  enum: [app, miniProgram]
                  default: app
      responses:
        "200":
          description: Prepay params (mock=true on staging when using mock provider)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /orders/{id}/refund:
    post:
      tags: [orders]
      summary: Create refund skeleton
      operationId: refundOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Refund transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /admin/account-deletions/{id}/orders/{orderId}/payment/sync:
    post:
      tags: [admin]
      summary: Synchronize a deletion-order payment
      operationId: adminSyncDeletionPayment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Payment synchronized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /admin/account-deletions/{id}/orders/{orderId}/refund/sync:
    post:
      tags: [admin]
      summary: Synchronize a deletion-order refund
      operationId: adminSyncDeletionRefund
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Refund synchronized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Envelope error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Meta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
    SuccessEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/Meta"
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
        meta:
          $ref: "#/components/schemas/Meta"
...
    CompanionTodayServiceEntry:
      type: object
      required: [id, scheduledAt, durationMinutes, status, serviceTitle]
      description: Minimal owner-only workbench item; it intentionally contains no customer, conversation, refund, settlement, payment, or internal order fields.
      properties:
        id: { type: string }
        scheduledAt: { type: string, format: date-time }
        durationMinutes: { type: integer, minimum: 1 }
        status: { type: string, enum: [pending, paying, paid, inService, completed] }
        serviceTitle: { type: string }
    CompanionTodayServiceSchedule:
      type: object
      required: [date, timezone, pendingConfirmationCount, items]
      description: Read-only list for one Asia/Shanghai calendar day. The date boundary is computed by the server from appointment start times.
      properties:
        date: { type: string, format: date, example: "2026-07-21" }
        timezone: { type: string, enum: [Asia/Shanghai] }
        pendingConfirmationCount:
          type: integer
          minimum: 0
          description: Caller-owned pending orders that still await companion confirmation; the count does not create a reminder or change an order.
        items:
          type: array
          items:
            $ref: "#/components/schemas/CompanionTodayServiceEntry"
    OrderRescheduleRequest:
      type: object
      required: [id, requestedByRole, originalScheduledAt, requestedScheduledAt, requestedAvailabilitySnapshot, status, expiresAt, respondedAt]
      properties:
        id: { type: string }
        requestedByRole: { type: string, enum: [customer, companion] }
        originalScheduledAt: { type: string, format: date-time }
        requestedScheduledAt: { type: string, format: date-time }
        requestedAvailabilitySnapshot:
          type: object
          nullable: true
          required: [availabilityWindowId, startsAt, endsAt, capacity]
          properties:
            availabilityWindowId: { type: string }
            startsAt: { type: string, format: date-time, nullable: true }
            endsAt: { type: string, format: date-time, nullable: true }
            capacity: { type: integer, nullable: true, minimum: 1 }
        status: { type: string, enum: [pending, accepted, rejected, expired, cancelled] }
        expiresAt: { type: string, format: date-time }
        respondedAt: { type: string, format: date-time, nullable: true }
    OrderTimelineEvent:
      type: object
      required: [id, type, actorRole, occurredAt, rescheduleRequest]
      properties:
        id: { type: string }
        type: { type: string, enum: [orderCreated, rescheduleRequested, rescheduleAccepted, rescheduleRejected, rescheduleExpired, rescheduleCancelled] }
        actorRole: { type: string, enum: [customer, companion, system] }
        occurredAt: { type: string, format: date-time }
        rescheduleRequest:
          type: object
          nullable: true
          allOf:
            - $ref: "#/components/schemas/OrderRescheduleRequest"
    VoiceRoomAccess:
      type: object
      required: [provider, sdkAppId, roomId, userId, userSig, privateMapKey, participantRole, expiresAt, serviceEndsAt, participant]
      description: Short-lived connection material. It is valid only for the caller, room and service window returned by the server; clients must keep it in memory only.
      properties:
        provider: { type: string, enum: [trtc] }
        sdkAppId: { type: integer, minimum: 1 }
        roomId: { type: string, maxLength: 64, description: Opaque provider room identifier derived by the server. }
        userId: { type: string, maxLength: 32, description: Opaque provider participant identifier; not the platform account ID. }
        userSig: { type: string, minLength: 1, description: Short-lived provider credential; never persist or log. }
        privateMapKey: { type: string, minLength: 1, description: Provider room restriction for this participant; never persist or log. }
        participantRole: { type: string, enum: [customer, companion] }
        expiresAt: { type: string, format: date-time }
        serviceEndsAt: { type: string, format: date-time }
        participant:
          type: object
          required: [name, initials]
          properties:
            name: { type: string, maxLength: 80 }
            initials: { type: string, maxLength: 16 }
...